move credential audit log request tests to shared examples
